+++ scrooge-core/src/main/scala/com/twitter/scrooge/StructBuilder.scala
@@ -1,6 +1,7 @@
 package com.twitter.scrooge
 
 import com.twitter.util.Memoize
+import org.apache.thrift.protocol.TType
 import scala.reflect.ClassTag
 
 /**
@@ -8,9 +9,12 @@ import scala.reflect.ClassTag
  * ThriftStruct or statically from T.
  *
  * We pass in a list of [[ClassTag]]s which describe each of the struct's field types
- * so that we can validate the values we are setting at runtime.
+ * so that we can validate the values we are setting at runtime. We also pass in a set
+ * of indices of non-primitive default-required fields; those can have null as valid value.
  */
-abstract class StructBuilder[T <: ThriftStruct](fieldTypes: IndexedSeq[ClassTag[_]]) {
+abstract class StructBuilder[T <: ThriftStruct](
+  fieldTypes: IndexedSeq[ClassTag[_]],
+  nullableIndices: Set[Int]) {
   protected val fieldArray: Array[Any] = new Array[Any](fieldTypes.size)
 
   /**
@@ -23,6 +27,7 @@ abstract class StructBuilder[T <: ThriftStruct](fieldTypes: IndexedSeq[ClassTag[
   private[this] def addOrUpdateFieldArray[A](index: Int, v: Any)(implicit tag: ClassTag[A]): Unit =
     v match {
       case inputValue: A => fieldArray(index) = inputValue
+      case null if nullableIndices(index) => fieldArray(index) = StructBuilder.ExplicitNull
       case _ => throw new IllegalArgumentException(s"value at index $index must be of type $tag")
     }
 
@@ -84,10 +89,42 @@ abstract class StructBuilder[T <: ThriftStruct](fieldTypes: IndexedSeq[ClassTag[
  * This object provides operations to obtain `StructBuilder` instances.
  */
 object StructBuilder {
+  // Used as an explicitly set null value for default-required fields
+  private[StructBuilder] object ExplicitNull
+
+  def unwrapExplicitNull(v: Any): Any =
+    v match {
+      case ExplicitNull => null
+      case _ => v
+    }
+
   private[this] val memoizeBuilderMethod: Class[_] => () => Any = Memoize.classValue { clazz =>
     val thriftCodec = ThriftStructCodec.forStructClass(clazz.asSubclass(classOf[ThriftStruct]))
     val m = thriftCodec.getClass.getMethod("newBuilder")
     () => m.invoke(thriftCodec)
+  }
+
+  /**
+   * Given a sequence of struct field information, returns a set of indices of non-primitive
+   * default-required fields.
+   * @param fieldInfos a sequence of field informations
+   * @return a set of indices for non-primitive default-required fields
+   */
+  final def nullableIndices(fieldInfos: Seq[ThriftStructFieldInfo]): Set[Int] = {
+    fieldInfos.zipWithIndex.collect {
+      case (f, i) if isNullable(f) => i
+    }.toSet
+  }
+
+  private[this] def isNullable(f: ThriftStructFieldInfo): Boolean = {
+    if (f.isOptional || f.isRequired) {
+      false
+    } else {
+      f.tfield.`type` match {
+        case TType.BOOL | TType.BYTE | TType.I16 | TType.I32 | TType.I64 | TType.DOUBLE => false
+        case _ => true
+      }
+    }
   }
 
   /**