+++ scrooge-generator-tests/src/test/resources/gold_file_output_scala/com/twitter/scrooge/test/gold/thriftscala/GoldService.scala
@@ -329,6 +329,8 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
       )
     
     
+      lazy val nullableIndices: Set[Int] = StructBuilder.nullableIndices(fieldInfos)
+    
       val structAnnotations: immutable$Map[String, String] =
         immutable$Map.empty[String, String]
     
@@ -425,7 +427,7 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
         )
       }
     
-      def newBuilder(): StructBuilder[Args] = new ArgsStructBuilder(_root_.scala.None, fieldTypes)
+      def newBuilder(): StructBuilder[Args] = new ArgsStructBuilder(_root_.scala.None, fieldTypes, nullableIndices)
     
       override def encode(_item: Args, _oproto: TProtocol): Unit = {
         _item.write(_oproto)
@@ -589,27 +591,27 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
     
       def _codec: ValidatingThriftStructCodec3[Args] = Args
     
-      def newBuilder(): StructBuilder[Args] = new ArgsStructBuilder(_root_.scala.Some(this), fieldTypes)
+      def newBuilder(): StructBuilder[Args] = new ArgsStructBuilder(_root_.scala.Some(this), fieldTypes, nullableIndices)
     }
     
-    private[thriftscala] class ArgsStructBuilder(instance: _root_.scala.Option[Args], fieldTypes: IndexedSeq[ClassTag[_]])
-        extends StructBuilder[Args](fieldTypes) {
+    private[thriftscala] class ArgsStructBuilder(instance: _root_.scala.Option[Args], fieldTypes: IndexedSeq[ClassTag[_]], nullableIndices: Set[Int])
+        extends StructBuilder[Args](fieldTypes, nullableIndices) {
     
       def build(): Args = {
         val _fieldArray = fieldArray // shadow variable
         if (instance.isDefined) {
           val instanceValue = instance.get
           Args(
-            if (_fieldArray(0) == null) instanceValue.request else _fieldArray(0).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.Request],
-            if (_fieldArray(1) == null) instanceValue.unionRequest else _fieldArray(1).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.RequestUnion],
-            if (_fieldArray(2) == null) instanceValue.exceptionRequest else _fieldArray(2).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.RequestException]
+            if (_fieldArray(0) == null) instanceValue.request else StructBuilder.unwrapExplicitNull(_fieldArray(0)).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.Request],
+            if (_fieldArray(1) == null) instanceValue.unionRequest else StructBuilder.unwrapExplicitNull(_fieldArray(1)).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.RequestUnion],
+            if (_fieldArray(2) == null) instanceValue.exceptionRequest else StructBuilder.unwrapExplicitNull(_fieldArray(2)).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.RequestException]
           )
         } else {
           if (genericArrayOps(_fieldArray).contains(null)) throw new InvalidFieldsException(structBuildError("Args"))
           Args(
-            _fieldArray(0).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.Request],
-            _fieldArray(1).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.RequestUnion],
-            _fieldArray(2).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.RequestException]
+            StructBuilder.unwrapExplicitNull(_fieldArray(0)).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.Request],
+            StructBuilder.unwrapExplicitNull(_fieldArray(1)).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.RequestUnion],
+            StructBuilder.unwrapExplicitNull(_fieldArray(2)).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.RequestException]
           )
         }
       }
@@ -654,6 +656,8 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
         )
       )
     
+    
+      lazy val nullableIndices: Set[Int] = StructBuilder.nullableIndices(fieldInfos)
     
       val structAnnotations: immutable$Map[String, String] =
         immutable$Map.empty[String, String]
@@ -739,7 +743,7 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
         )
       }
     
-      def newBuilder(): StructBuilder[Result] = new ResultStructBuilder(_root_.scala.None, fieldTypes)
+      def newBuilder(): StructBuilder[Result] = new ResultStructBuilder(_root_.scala.None, fieldTypes, nullableIndices)
     
       override def encode(_item: Result, _oproto: TProtocol): Unit = {
         _item.write(_oproto)
@@ -886,11 +890,11 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
     
       def _codec: ValidatingThriftStructCodec3[Result] = Result
     
-      def newBuilder(): StructBuilder[Result] = new ResultStructBuilder(_root_.scala.Some(this), fieldTypes)
+      def newBuilder(): StructBuilder[Result] = new ResultStructBuilder(_root_.scala.Some(this), fieldTypes, nullableIndices)
     }
     
-    private[thriftscala] class ResultStructBuilder(instance: _root_.scala.Option[Result], fieldTypes: IndexedSeq[ClassTag[_]])
-        extends StructBuilder[Result](fieldTypes) {
+    private[thriftscala] class ResultStructBuilder(instance: _root_.scala.Option[Result], fieldTypes: IndexedSeq[ClassTag[_]], nullableIndices: Set[Int])
+        extends StructBuilder[Result](fieldTypes, nullableIndices) {
     
       def build(): Result = {
         val _fieldArray = fieldArray // shadow variable
@@ -971,6 +975,8 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
       )
     
     
+      lazy val nullableIndices: Set[Int] = StructBuilder.nullableIndices(fieldInfos)
+    
       val structAnnotations: immutable$Map[String, String] =
         immutable$Map.empty[String, String]
     
@@ -1035,7 +1041,7 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
         )
       }
     
-      def newBuilder(): StructBuilder[Args] = new ArgsStructBuilder(_root_.scala.None, fieldTypes)
+      def newBuilder(): StructBuilder[Args] = new ArgsStructBuilder(_root_.scala.None, fieldTypes, nullableIndices)
     
       override def encode(_item: Args, _oproto: TProtocol): Unit = {
         _item.write(_oproto)
@@ -1156,23 +1162,23 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
     
       def _codec: ValidatingThriftStructCodec3[Args] = Args
     
-      def newBuilder(): StructBuilder[Args] = new ArgsStructBuilder(_root_.scala.Some(this), fieldTypes)
+      def newBuilder(): StructBuilder[Args] = new ArgsStructBuilder(_root_.scala.Some(this), fieldTypes, nullableIndices)
     }
     
-    private[thriftscala] class ArgsStructBuilder(instance: _root_.scala.Option[Args], fieldTypes: IndexedSeq[ClassTag[_]])
-        extends StructBuilder[Args](fieldTypes) {
+    private[thriftscala] class ArgsStructBuilder(instance: _root_.scala.Option[Args], fieldTypes: IndexedSeq[ClassTag[_]], nullableIndices: Set[Int])
+        extends StructBuilder[Args](fieldTypes, nullableIndices) {
     
       def build(): Args = {
         val _fieldArray = fieldArray // shadow variable
         if (instance.isDefined) {
           val instanceValue = instance.get
           Args(
-            if (_fieldArray(0) == null) instanceValue.request else _fieldArray(0).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.Request]
+            if (_fieldArray(0) == null) instanceValue.request else StructBuilder.unwrapExplicitNull(_fieldArray(0)).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.Request]
           )
         } else {
           if (genericArrayOps(_fieldArray).contains(null)) throw new InvalidFieldsException(structBuildError("Args"))
           Args(
-            _fieldArray(0).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.Request]
+            StructBuilder.unwrapExplicitNull(_fieldArray(0)).asInstanceOf[com.twitter.scrooge.test.gold.thriftscala.Request]
           )
         }
       }
@@ -1203,6 +1209,8 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
         )
       )
     
+    
+      lazy val nullableIndices: Set[Int] = StructBuilder.nullableIndices(fieldInfos)
     
       val structAnnotations: immutable$Map[String, String] =
         immutable$Map.empty[String, String]
@@ -1270,7 +1278,7 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
         )
       }
     
-      def newBuilder(): StructBuilder[Result] = new ResultStructBuilder(_root_.scala.None, fieldTypes)
+      def newBuilder(): StructBuilder[Result] = new ResultStructBuilder(_root_.scala.None, fieldTypes, nullableIndices)
     
       override def encode(_item: Result, _oproto: TProtocol): Unit = {
         _item.write(_oproto)
@@ -1393,11 +1401,11 @@ object GoldService extends _root_.com.twitter.finagle.thrift.GeneratedThriftServ
     
       def _codec: ValidatingThriftStructCodec3[Result] = Result
     
-      def newBuilder(): StructBuilder[Result] = new ResultStructBuilder(_root_.scala.Some(this), fieldTypes)
+      def newBuilder(): StructBuilder[Result] = new ResultStructBuilder(_root_.scala.Some(this), fieldTypes, nullableIndices)
     }
     
-    private[thriftscala] class ResultStructBuilder(instance: _root_.scala.Option[Result], fieldTypes: IndexedSeq[ClassTag[_]])
-        extends StructBuilder[Result](fieldTypes) {
+    private[thriftscala] class ResultStructBuilder(instance: _root_.scala.Option[Result], fieldTypes: IndexedSeq[ClassTag[_]], nullableIndices: Set[Int])
+        extends StructBuilder[Result](fieldTypes, nullableIndices) {
     
       def build(): Result = {
         val _fieldArray = fieldArray // shadow variable